Initial changes to support taxes
