+++ Dockerfile
@@ -1,5 +1,7 @@
 FROM node:18.18.0-alpine
 RUN mkdir -p /tmp/actual
+COPY app/base-taxes.yml /tmp/actual/
+RUN mkdir -p /usr/src/tax
 WORKDIR /usr/src/app
 COPY package*.json ./
 RUN apk add -U tzdata
