+++ app/base-taxes.yml
@@ -1,0 +1,1 @@
+# empty file
